requirements.txt
fastapi
uvicorn
langchain-community>=0.4.0
langchain-core
langchain
langchain-google-genai
langchain-huggingface
sentence-transformers
langchain-classic
langchain-pinecone==0.2.13
pinecone
pypdf
python-dotenv
pydantic
python-multipart
certifi
sqlalchemy
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
google-generativeai
